move check to config module and clean up imports
